requirements.txt
diffusers>=0.33.0
accelerate
transformers
einops
torchvision
protobuf
sentencepiece